Load image on click in source file tree
I don't really like this though. I think I'm going to subclass the
FileTreePanel instead because I want different behavior between the two
trees, which I could achive with this, but it feels gross to have to
call setFileTreePanel on the listener with the circular dependency. It'd
probably be cleaner to push it into the class itself instead.